pyproject.toml
[project]
name = "pioneerxity-web-scraper-service"
version = "0.1.0"
description = "Internal gRPC microservice for documentation scraping, filtering, document generation, and MinIO storage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "grpcio>=1.58.0",
  "grpcio-tools>=1.58.0",
  "protobuf>=4.25.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "python-dotenv>=1.0.0",
  "structlog>=23.2.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "minio>=7.2.0",
  "playwright>=1.40.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "aiohttp>=3.9.0",
  "weasyprint>=60.0",
  "reportlab>=4.0.0",
  "python-docx>=1.1.0",
  "pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "ruff>=0.1.7",
]

[tool.uv]
dev-dependencies = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "ruff>=0.1.7",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper"]



